Remove Flutter sample.
This was added 3 years ago as an experiment.
Unfortunately, this was never continued, and now it's completely abandoned and not supported (CI was disabled couple of years ago because it was failing due to broken dependencies.
The decision was made to not continue with it.
Thus I'm removing it from the samples.

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>