C++ code: eliminate a source of potential bugs
To create zero-filled structs, we use a method where we create
a static struct using a 'dummy constructor' taking an (unused) int argument.
The problem is that if you mistakenly have e.g.
    COPROC c = 1;
you don't get a compile error;
you get an uninitialized struct as if you had done
    COPROC c = COPROC(1);
I changed this so that we use a type DUMMY_TYPE
(guaranteed not to be convertible from anything)
as the arg type for the dummy constructors.
So now you get a compile error in the above case.