validator: handle transient errors
A 'transient error' is one that will go away in a while,
e.g. an fopen failure because of a broken NSF mount.
In general, the BOINC back-end code (validation, assimilation)
handles transient errors sensibly:
if there's a bad NSF mount, it retries validation for a few hours
rather than marking thousands of jobs as failed.

Extend this to script-based validation.
If a script (either init_result or compare_results) exits with 3,
treat that as a transient error.
Treat other nonzero exits (or lack of an exit code) as a permanent error.

More generally (for all validators) add a return value
VAL_RESULT_TRANSIENT_ERROR for init_result() and compute_results().
This means any transient error.
Previously we checked only for ERR_OPENDIR.
And for compare_results() we treated all nonzero returns as permanent.