[Windows] Set correct platform name when running BOINC client on Windows on ARM.
With this commit a new platform name 'windows_arm64' added.
Also, according to the https://learn.microsoft.com/en-us/windows/arm/apps-on-arm-x86-emulation, Windows 10 on ARM supports x86 applications, and Windows 11 on ARM supports x64 applications.
That is why we will be checking what version of Windows we are running on, and adding 'windows_intelx86' on Windows 10 and later (I suppose, this is will set always) and 'windows_x86_64' on Windows 11 and later (nothing later at the moment, but we are looking into the future!).

This fixes #5934.

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>