vboxwrapper: actually create the 'virtualbox home directory'
As far as I can tell this dir is used only on Win;
we run VBoxSVC.exe there, and look for a log file later.
Should we remove it for other platforms?

Also: there's some code to create a 'scratch' dir, projects/scratch.
This seems like a bad idea; we shouldn't put random stuff in projects/,
and also if there are multiple VM jobs they share the same dir.
Should we get rid of this?