Fix snprintf() errors introduced in PR #4864
PR #4864 introduced 4 erroneous changes of the form
   sprintf(buf, sizeof(buf), ....)
where buf is a char*.
(Note: in this case sizeof(buf) is 8.
It is NOT the size of the memory area that buf points to).

One of these (in file_names.cpp) was fixed in a later PR.
This PR fixes the other 3.
Hopefully we didn't miss any.