client and docker_wrapper (Mac/podman): set env vars to use accessable dir.
Podman on Mac was failing because it tried to store data in
.config/ in the user's home dir.
Podman runs as boinc_master (in the client)
and boinc_projects (in the wrapper)
and neither account has access to this dir.

To fix this, we'll have the installer create an accessable dir
'podman' in the BOINC data dir.
This change prepends an 'env' directive to Podman commands on Mac,
telling it to use this dir.