Server: fix DB purge for batch-oriented apps
The BOINC back-end architecture was originally designed
for stream-oriented apps, in which jobs are processed, assimilated,
file-deleted, and then purged from the DB.

Recently we've added support for batch-oriented apps,
for which we don't want to purge the WU and result records
until the batch has been retired.

Fix: add an option --retired_wus to db_purge.
If set, it purges only WUs and results in retired batches.
(If not set, it purges only WUs that are not in a batch;
so if your project has both stream- and batch-oriented apps,
run one of each).

Also: fix the logic for retire_batch().
It shouldn't set WUs to FILE_DELETE_DONE.
WUs may be in progress, and we still need to file-delete them.
Instead, mark unsent WUs as WU_ERROR_CANCELLED so we don't send them.

Also: add a script ops/delete_batches.php which deletes
retired batches with no WUs (i.e. they've all been purged).
This improves the efficiency of db_purge.
You don't need to do it very often - once a week is enough.

Also: aborting a batch sets its WU error masks to CANCELLED,
in which case the WUs won't go through file-deletion
and their input files will be orphaned.
Fix: when aborting a job, set the WU file delete state to READY.
Similar for unsent jobs when you retire a batch.