web: app-level access control for BUDA
Each BUDA app has a creator, and also a list of 'submitter' user IDs.

Only the creator can edit the app, or create or edit variants.
The creator can add or remove submitters
(but they have to have submit access to BUDA as a whole).

Only the creator and submitters can submit jobs to the app.