Client: use 'abort' rather than 'quit' when appropriate
The client has two ways to tell a job to exit:
'quit' if we plan to restart the job later;
'abort' if we're not going to restart it.

We were incorrectly using quit instead of abort
when detaching or resetting a project.

This doesn't make much difference in general, but for Docker jobs it does:
abort should remove the container and image, quit should not.

Also: rename ACTIVE_TASK::request_exit() to request_quit()
to clarify the distinction.